fix: update types and remove excess files
